Added configuration file and a logging service.
The logging service still needs to be able to write to disk, but that'll come when it's needed.

JSON is easy to bring in, as the parsing facilities are part of the standard lib, but it's not very forgiving syntax wise nor does it have nice automatic connections between variables and keys.  Ah well, works for now.