Added ability to connect to database to get regions and then download them.
Included is a new configuration parameter "database_connection" which is required if the --gettiles flag is set.

Note that the tile downloading process is SLOW: the regions take about 1/2 a second to create the tile and send it over.  Anything to reduce the number of calls or improve the speed is important here!