Exiting early from build is cleaner. LGRP_APP should be private to the module.
Some libraries can depend on other libs, and trying to handle that via the needsextract flag was getting bad - it was overloading the purpose and meaning, and having a flag in the first place is bad enough, adding more wasn't a good idea to me.
However the exit command has the /B flag which allow it to work like exit in UNIX shells: just exit the script, not the whole world.  This is what I wanted in the first place!

There's no need for any other module to see that constant.