Added ability to create the ocean tile.
The build system for this commit leaves much to be desired: Windows requires a ton of intervention, Linux is untouched, and the DMagick lib requires custom changes.  I'll be addressing this ASAP.

The ocean tile creation is a huge leap forward: now the image processing library is in place and operational - even if how to get there isn't obvious at this stage.  I'll get that ironed out and the actual zoom algorithm built shortly.

Not entirely sure why the tile color isn't exactly as requested, but for now it is close enough.