Build system improvements.
More verbosity during time-consuming activities.  Makes the time not feel so long.
Cleaned out non-essential actions, such as building the DMagick library - we are using the source directly!

Temporarily swapped to using my fork of DMagick to get support for the DMagick_No_Display version flag.  This allows us to avoid having to link with User32.lib and waste space and time on GUI functionality that will never be used.