Fixed Issue 5: Review ATileGrabber class for collapse to function.
Turns out that it could be collapsed, but it still needed to be separate functions: the list from the DB will be needed for the zoom operation.

I also discovered that I'd missed an assumption violation: originally the config file was purely optional, and the code treaded it as such.  However, as soon as a required entry is put in place the config file is no longer optional.
And now that the database connection is required to perform the core task, that config file is definitely front-and-center.