Added code to gather the correct region tiles into the temp folder.
There's still a possibility that there could be tiles that are not in the DB (from the new tiles folder).  And if the DB got a region added between the last filling of the new tile folder and a call to this program (without the --getttiles flag), there could be missed tiles.
Both potential issues are minor, as the next cycle should cause a correction.