Handled a null ref crash.
When a region is in the DB / marked as online, --gettiles is NOT specified, there's no image for that region in neither the maptiles nor the newtiles folder, there would wind up being a null ref in the compositing causing a crash.

Handled by giving that region a bright red solid fill to indicate that it is missing.  Under normal operations the region should have sent a map tile to the newtiles folder upon startup, so not having a tile means that something's not right and it shoudl be very obvious.